project1/pyproject.toml
[project]
name = "project1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-ollama>=0.3.3",
    "langchain-openai>=0.3.27",
    "langgraph>=0.5.1",
    "python-dotenv>=1.1.1",
]
